+++ ic-agent/src/agent/replica_api.rs
@@ -1,7 +1,5 @@
-use crate::{
-    export::Principal,
-    hash_tree::{HashTree, Label},
-};
+use crate::export::Principal;
+use ic_types::Label;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -104,29 +102,6 @@ pub enum SyncContent {
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub struct ReadStateResponse {
-    #[serde(with = "serde_bytes")]
-    pub certificate: Vec<u8>,
-}
-
-/// A `Certificate` as defined in <https://smartcontracts.org/docs/interface-spec/index.html#_certificate>
-#[derive(Debug, Deserialize, PartialEq, Eq)]
-pub struct Certificate<'a> {
-    /// The hash tree.
-    pub tree: HashTree<'a>,
-
-    /// The signature of the root hash in `tree`.
-    #[serde(with = "serde_bytes")]
-    pub signature: Vec<u8>,
-
-    /// A delegation from the root key to the key used to sign `signature`, if one exists.
-    pub delegation: Option<Delegation>,
-}
-
-#[derive(Debug, Deserialize, PartialEq, Eq)]
-pub struct Delegation {
-    #[serde(with = "serde_bytes")]
-    pub subnet_id: Vec<u8>,
-
     #[serde(with = "serde_bytes")]
     pub certificate: Vec<u8>,
 }