+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.22.0] - 2022-10-17
 
+* Create `ic-types` crate and move HashTree and Certificate types.
 * Drop `disable_range_check` flag from certificate delegation checking.
 
 ## [0.21.0] - 2022-10-03